Rename to socket_stream
